speed_score/requirement.txt
appdirs==1.4.4
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
configparser==5.0.2
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
distlib==0.3.1
et-xmlfile==1.0.1
filelock==3.0.12
hyperlink==21.0.0
idna==3.1
importlib-metadata==3.7.2
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jdcal==1.4.1
jmespath==0.10.0
lxml==4.6.2
mysql-client==0.0.1
mysql-connector-python==8.0.23
packaging==20.9
parsel==1.6.0
Pillow==8.1.2
pip-review==1.1.0
Protego==0.1.16
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
queuelib==1.5.0
reportlab==3.5.65
requests==2.25.1
Scrapy==2.4.1
scrapy-user-agents==0.1.1
scrapyd==1.2.1
scrapyd-client==1.1.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.2
sriram-proxy-pool==0.2.2
sriram-proxyscrape==0.4.0
stem==1.8.0
Twisted==21.2.0
typing-extensions==3.7.4.3
ua-parser==0.10.0
urllib3==1.26.3
user-agents==2.2.0
virtualenv==20.4.2
w3lib==1.22.0
zipp==3.4.1
zope.interface==5.2.0
